% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create3dGenomicSignals.R
\name{create3dGenomicSignals}
\alias{create3dGenomicSignals}
\title{create 3d Geometry by given genomic signals}
\usage{
create3dGenomicSignals(
  GenoSig,
  targetObj,
  signalTransformFun,
  reverseGenomicSigs,
  type = "segment",
  tag,
  name,
  color = c("gray30", "darkred"),
  ...
)
}
\arguments{
\item{GenoSig}{The Genomic signals. An object of 
\link[GenomicRanges:GRanges-class]{GRanges} with scores or
 an object of \link[trackViewer:track]{track}.}

\item{targetObj}{The GRanges object with mcols x0, y0, z0, x1, y1, and z1}

\item{signalTransformFun}{The transformation function for genomic signals.}

\item{reverseGenomicSigs}{Plot the genomic signals in reverse values.}

\item{type}{The Geometry type.See \link{threeJsGeometry}}

\item{tag}{The tag used to group geometries.}

\item{name}{The prefix for the name of the geometries.}

\item{color}{The color of the signal.}

\item{...}{the parameters for each different type of geometries.
If type is 'segments', lwd.maxGenomicSigs (the maximal lwd of the line) is
 required.
If type is 'circle', radius (the radius of the circle) and 
 the maxVal (the value for 2*pi) is
 required.}
}
\value{
\link{threeJsGeometry} objects or NULL
}
\description{
Create a 3d Geometry by given genomic signals for target 3d positions.
}
