% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClusters.R
\name{fill_NA}
\alias{fill_NA}
\title{fill NA values by upstream and downstream points}
\usage{
fill_NA(xyz)
}
\arguments{
\item{xyz}{A matrix or data.frame with columns 'x', 'y', 'z'}
}
\value{
A matrix or data.frame.
}
\description{
Fill NA values by previous and next points coordinates.
}
\examples{
xyz <- matrix(seq.int(21), ncol=3, dimnames=list(NULL, c('x', 'y', 'z')))
xyz[c(1, 5, 7), ] <- NA
fill_NA(xyz)

}
