% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDistanceMatrix.R
\name{gaussianBlur}
\alias{gaussianBlur}
\title{Gaussian blur}
\usage{
gaussianBlur(mat, size = 5, sigma = 1, ...)
}
\arguments{
\item{mat}{A matrix.}

\item{size}{The kernel size}

\item{sigma}{The strength of the blur.}

\item{...}{Not used.}
}
\value{
A matrix.
}
\description{
Do Gaussian for the distance matrix.
}
\examples{
mat <- matrix(runif(100), 10, 10)
blurred_mat <- gaussianBlur(mat, size = 5, sigma = 1)
}
