% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClusters.R
\name{SDC}
\alias{SDC}
\title{Distance to centroid}
\usage{
SDC(xyz)
}
\arguments{
\item{xyz}{A data.frame with x, y, z coordinates.}
}
\value{
The mean of squared Euclidean distance to the centroid.
}
\description{
Calculates the mean of distance from each point to the geometric center (centroid)
}
\examples{
xyz <- matrix(seq.int(12), ncol = 3, dimnames=list(NULL, c('x', 'y', 'z')))
SDC(xyz)
}
