% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth3dPoints.R
\name{smooth3dPoints}
\alias{smooth3dPoints}
\title{Calculate the smoothed curve for input GRanges}
\usage{
smooth3dPoints(obj, resolusion = 30, ...)
}
\arguments{
\item{obj}{GRanges object with mcols x, y, and z}

\item{resolusion}{number of points at which to evaluate the smooth curve.}

\item{...}{parameters passed to \link[stats]{splinefun}}
}
\value{
GRanges object with smoothed points of x, y, and z.
}
\description{
Calculate the smoothed curve for input GRanges
}
\examples{
library(GenomicRanges)
obj <- GRanges('1', IRanges(seq.int(5)*10, width=10),
 x=seq.int(5), y=seq.int(5), z=seq.int(5))
smooth3dPoints(obj, 5)

}
