% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeJsGeometry.R
\docType{class}
\name{threeJsGeometry-class}
\alias{threeJsGeometry-class}
\alias{threeJsGeometry}
\alias{$,threeJsGeometry-method}
\alias{$<-,threeJsGeometry-method}
\title{Class \code{"threeJsGeometry"}}
\usage{
threeJsGeometry(...)

\S4method{$}{threeJsGeometry}(x, name)

\S4method{$}{threeJsGeometry}(x, name) <- value
}
\arguments{
\item{\dots}{Each argument in \dots becomes an slot in the new threeJsGeometry.}

\item{x}{an object of threeJsGeometry}

\item{name}{slot name of threeJsGeometry}

\item{value}{value to be assigned}
}
\description{
An object of class \code{"threeJsGeometry"} 
             represents `three.js` geometry.
}
\section{Slots}{

\describe{
\item{\code{x,y,z}}{\code{"numeric"}, specify the x, y, and z coordinates.}

\item{\code{colors}}{\code{"character"}, the colors for each geometry.}

\item{\code{type}}{\code{"charater"}, the type of the geometry. Available types are
'line', 'box', 'capsule', 'cone', 'cylinder', 'dodecahedron', 'icosahedron',
'octahedron', 'sphere', 'tetrahedron', 'text', and 'torus'.}

\item{\code{properties}}{A \code{"list"}, the properties to control the geometry.}
}}

\examples{
tjg <- threeJsGeometry()

}
