% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeJsGeometry.R
\name{availableGeometries}
\alias{availableGeometries}
\title{Available Geometries}
\usage{
availableGeometries
}
\description{
The Geometries supported by \link{threeJsGeometry} class
}
\examples{
availableGeometries

}
