% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeJsGeometry.R
\docType{data}
\name{availableGeometries}
\alias{availableGeometries}
\title{Available Geometries}
\format{
An object of class \code{character} of length 18.
}
\usage{
availableGeometries
}
\description{
The Geometries suported by \link{threeJsGeometry} class
}
\examples{
availableGeometries

}
\keyword{datasets}
