% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClusters.R
\name{cellClusters}
\alias{cellClusters}
\alias{cellDistance}
\title{cluster single cell 3D structures}
\usage{
cellClusters(
  xyzs,
  TADs,
  method = "ward.D2",
  quite = FALSE,
  parallel = FALSE,
  ...
)

cellDistance(xyzs, TADs, quite = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{xyzs}{A data.frame with x, y, z coordinates or output of cellDistance.}

\item{TADs}{A list of index vectors, where each vector represents a TAD.
For example, if the first TAD spans the 2nd to 4th coordinates and the
 second spans the 8th to 10th coordinates, the list would be:
  list(c(2, 3, 4), c(8, 9, 10)).}

\item{method}{The agglomeration method to be used for \link{hclust}.
Default is 'ward.D2'.}

\item{quite}{Print the message or not.}

\item{parallel}{Run parallel by future or not.}

\item{...}{not used.}
}
\value{
cellClusters return an object of class hclust.

cellDistance return distance matrix as an object of 'dist'
}
\description{
Perform Hierarchical clustering for given 3D structures.

Calculate euclidean distance for each pair of cells after alignment.
}
\examples{
set.seed(1)
xyzs <- lapply(seq.int(20), function(i){
  matrix(sample.int(100, 60, replace = TRUE),
   nrow=20, dimnames=list(NULL, c('x', 'y', 'z')))
})
cd <- cellDistance(xyzs)
cc <- cellClusters(cd)
# plot(cc)
cutree(cc, k=3)
}
