% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglViewer.R
\name{rglViewer}
\alias{rglViewer}
\title{rgl Viewer
View the 3d structure by rgl.}
\usage{
rglViewer(..., background = "#00000088")
}
\arguments{
\item{...}{objects of threeJsGeometry.}

\item{background}{background of the main camera.}
}
\value{
MULL
}
\description{
rgl Viewer
View the 3d structure by rgl.
}
\examples{
# example code
}
